chore(release): defer dcode status to v0.0.74
Revert 470060d2459ea04fdaeca0c00219c38d221ad247 so v0.0.73 retains its intended release boundary.

PR #6202 remains targeted for v0.0.74.
