+++ src/lib/onboard.ts
@@ -2998,9 +2998,10 @@ async function createSandbox(
   const envMessagingState = MessagingHostStateApplier.readPlanStateFromEnv();
   const plannedMessagingState =
     envMessagingState?.plan.sandboxName === sandboxName ? envMessagingState : undefined;
+  const plannedMessagingPlan = plannedMessagingState?.plan;
   sandboxBuildPatchConfig.prepareSandboxBuildPatchConfig({
     configuredMessagingChannels:
-      getChannelsFromPlan(plannedMessagingState?.plan) ?? activeMessagingChannels,
+      getChannelsFromPlan(plannedMessagingPlan) ?? activeMessagingChannels,
   });
   const { buildId } = await sandboxDockerfilePatchFlow.prepareSandboxDockerfilePatch({
     agent,
@@ -3024,7 +3025,6 @@ async function createSandbox(
       agent,
       chatUiUrl,
       createArgs,
-      sandboxName,
       env: process.env,
       extraPlaceholderKeys,
       getDashboardForwardPort,
