--- test/dcode-wrapper-identity.test.ts
@@ -1,426 +1,0 @@
-// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-import { spawnSync } from "node:child_process";
-import fs from "node:fs";
-import os from "node:os";
-import path from "node:path";
-
-import { describe, expect, it } from "vitest";
-
-const WRAPPER = path.join(
-  import.meta.dirname,
-  "..",
-  "agents",
-  "langchain-deepagents-code",
-  "dcode-wrapper.sh",
-);
-
-const canRun = process.platform === "linux";
-
-const SAMPLE_CONFIG = [
-  "# Generated by NemoClaw. This file contains no provider secrets.",
-  "# NemoClaw provider route: inference; upstream provider: nvidia-prod; API: openai-completions.",
-  "",
-  "[agents]",
-  'default = "backend-dev"',
-  'recent = "frontend-dev"',
-  "",
-  "[models]",
-  'default = "openai:demo-model"',
-  "",
-  "[models.providers.openai]",
-  'models = ["demo-model"]',
-  'base_url = "https://inference.local/v1"',
-  "enabled = true",
-  "",
-].join("\n");
-
-const OPAQUE = "Zx3Qw9Lp7Rt2Vn5Bd8Kf1Mh6Cg4Js0Ay";
-const CANONICAL_TLS_KEY_PATH = "/etc/openshell/tls/client/tls.key";
-
-type Fixture = { wrapperPath: string; ranMarker: string; envFile: string; configDir: string };
-
-function buildFixture(tempDir: string, configContent: string): Fixture {
-  const wrapperPath = path.join(tempDir, "dcode");
-  const ranMarker = path.join(tempDir, "dcode-ran");
-  const envFile = path.join(tempDir, ".env");
-  const configFile = path.join(tempDir, "config.toml");
-  const fixture = fs
-    .readFileSync(WRAPPER, "utf8")
-    .replace(
-      'readonly DEEPAGENTS_ENV_FILE="/sandbox/.deepagents/.env"',
-      `readonly DEEPAGENTS_ENV_FILE="${envFile}"`,
-    )
-    .replace(
-      'readonly DEEPAGENTS_CONFIG_FILE="/sandbox/.deepagents/config.toml"',
-      `readonly DEEPAGENTS_CONFIG_FILE="${configFile}"`,
-    )
-    .replace(
-      "exec python3 -m deepagents_code",
-      `touch "${ranMarker}"; echo dcode-stub-ran; exit 0; : python3 -m deepagents_code`,
-    );
-  fs.writeFileSync(envFile, "", "utf8");
-  fs.writeFileSync(configFile, configContent, "utf8");
-  fs.writeFileSync(wrapperPath, fixture, "utf8");
-  fs.chmodSync(wrapperPath, 0o755);
-  return { wrapperPath, ranMarker, envFile, configDir: tempDir };
-}
-
-function addAgentDir(fixture: Fixture, name: string): void {
-  fs.mkdirSync(path.join(fixture.configDir, name));
-}
-
-type Run = { status: number | null; stdout: string; stderr: string; launched: boolean };
-
-function runBashWrapper(fixture: Fixture, args: readonly string[], env: NodeJS.ProcessEnv): Run {
-  const result = spawnSync("bash", [fixture.wrapperPath, ...args], {
-    env: {
-      PATH: process.env.PATH ?? "/usr/bin:/bin",
-      HOME: path.dirname(fixture.wrapperPath),
-      ...env,
-    },
-    encoding: "utf8",
-    timeout: 10000,
-  });
-  return {
-    status: result.status,
-    stdout: result.stdout ?? "",
-    stderr: result.stderr ?? "",
-    launched: fs.existsSync(fixture.ranMarker),
-  };
-}
-
-function withTempDir(run: (dir: string) => void): void {
-  const dir = fs.mkdtempSync(path.join(os.tmpdir(), "nemoclaw-dcode-identity-"));
-  try {
-    run(dir);
-  } finally {
-    fs.rmSync(dir, { recursive: true, force: true });
-  }
-}
-
-describe.skipIf(!canRun)(
-  "agents/langchain-deepagents-code/dcode-wrapper.sh identity command",
-  () => {
-    for (const sub of ["status", "whoami", "identity"]) {
-      it(`'${sub}' reports the sandbox identity and does not launch dcode`, () => {
-        withTempDir((dir) => {
-          const fixture = buildFixture(dir, SAMPLE_CONFIG);
-          addAgentDir(fixture, "backend-dev");
-          const run = runBashWrapper(fixture, [sub], {
-            NEMOCLAW_SANDBOX_NAME: "dcode-demo",
-          });
-
-          expect(run.status).toBe(0);
-          expect(run.launched).toBe(false);
-          expect(run.stdout).toContain("Sandbox:  dcode-demo");
-          expect(run.stdout).toContain("Harness:  langchain-deepagents-code");
-          expect(run.stdout).toContain("Agent:    backend-dev");
-          expect(run.stdout).toContain("Route:    inference");
-          expect(run.stdout).toContain("Provider: nvidia-prod");
-          expect(run.stdout).toContain("Model:    openai:demo-model");
-          expect(run.stdout).toContain("Endpoint: https://inference.local/v1");
-          expect(run.stdout).toContain("Runtime:  Deep Agents Code (terminal)");
-        });
-      });
-    }
-
-    it("uses a valid recent dcode agent when the configured default is stale", () => {
-      withTempDir((dir) => {
-        const fixture = buildFixture(dir, SAMPLE_CONFIG);
-        addAgentDir(fixture, "frontend-dev");
-        const run = runBashWrapper(fixture, ["status"], {});
-
-        expect(run.status).toBe(0);
-        expect(run.stdout).toContain("Agent:    frontend-dev");
-      });
-    });
-
-    it("uses the upstream default agent when configured preferences are stale", () => {
-      withTempDir((dir) => {
-        const run = runBashWrapper(buildFixture(dir, SAMPLE_CONFIG), ["status"], {});
-
-        expect(run.status).toBe(0);
-        expect(run.stdout).toContain("Agent:    agent (default)");
-      });
-    });
-
-    it("ignores traversal-shaped agent preferences", () => {
-      withTempDir((dir) => {
-        const config = SAMPLE_CONFIG.replace('default = "backend-dev"', 'default = ".."');
-        const fixture = buildFixture(dir, config);
-        addAgentDir(fixture, "frontend-dev");
-        const run = runBashWrapper(fixture, ["status"], {});
-
-        expect(run.status).toBe(0);
-        expect(run.stdout).toContain("Agent:    frontend-dev");
-      });
-    });
-
-    it("ignores agent preferences that dcode cannot activate", () => {
-      withTempDir((dir) => {
-        for (const invalidName of [".hidden", "   "]) {
-          const config = SAMPLE_CONFIG.replace(
-            'default = "backend-dev"',
-            `default = "${invalidName}"`,
-          );
-          const fixture = buildFixture(dir, config);
-          addAgentDir(fixture, invalidName);
-          addAgentDir(fixture, "frontend-dev");
-          const run = runBashWrapper(fixture, ["status"], {});
-
-          expect(run.status).toBe(0);
-          expect(run.stdout).toContain("Agent:    frontend-dev");
-          expect(run.stdout).not.toContain(`Agent:    ${invalidName}`);
-          fs.rmSync(path.join(fixture.configDir, "frontend-dev"), { recursive: true });
-        }
-      });
-    });
-
-    it("does not write control characters from mutable identity metadata", () => {
-      withTempDir((dir) => {
-        const escape = "\u001b[31m";
-        const config = SAMPLE_CONFIG.replace(
-          'default = "openai:demo-model"',
-          `default = "openai:${escape}spoof"`,
-        )
-          .replace("upstream provider: nvidia-prod", `upstream provider: nvidia-prod${escape}`)
-          .replace('base_url = "https://inference.local/v1"', "");
-        const run = runBashWrapper(buildFixture(dir, config), ["status"], {
-          NEMOCLAW_SANDBOX_NAME: `demo${escape}`,
-          OPENAI_BASE_URL: `https://inference.local/${escape}`,
-        });
-
-        expect(run.status).toBe(0);
-        expect(run.stdout).not.toContain("\u001b");
-        expect(run.stdout).toContain("Sandbox:  unknown");
-        expect(run.stdout).not.toContain("Provider:");
-        expect(run.stdout).not.toContain("Model:");
-        expect(run.stdout).not.toContain("Endpoint:");
-
-        const unsafeConfigEndpoint = SAMPLE_CONFIG.replace(
-          "https://inference.local/v1",
-          `https://inference.local/${escape}`,
-        );
-        const configEndpointRun = runBashWrapper(
-          buildFixture(dir, unsafeConfigEndpoint),
-          ["status"],
-          { OPENAI_BASE_URL: "https://safe-fallback.example.test/v1" },
-        );
-
-        expect(configEndpointRun.status).toBe(0);
-        expect(configEndpointRun.stdout).not.toContain("safe-fallback.example.test");
-        expect(configEndpointRun.stdout).not.toContain("Endpoint:");
-      });
-    });
-
-    it("does not write oversized mutable identity metadata", () => {
-      withTempDir((dir) => {
-        const oversized = "x".repeat(257);
-        const config = SAMPLE_CONFIG.replace('base_url = "https://inference.local/v1"', "");
-        const run = runBashWrapper(buildFixture(dir, config), ["status"], {
-          NEMOCLAW_SANDBOX_NAME: oversized,
-          OPENAI_BASE_URL: oversized,
-        });
-
-        expect(run.status).toBe(0);
-        expect(run.stdout).toContain("Sandbox:  unknown");
-        expect(run.stdout).not.toContain(oversized);
-        expect(run.stdout).not.toContain("Endpoint:");
-      });
-    });
-
-    it("does not write secret-shaped mutable identity metadata", () => {
-      withTempDir((dir) => {
-        const agentSecret = "PASSWORD opaquevalue12345";
-        fs.mkdirSync(path.join(dir, agentSecret));
-        const secretValues = [
-          `tvly-${OPAQUE}`,
-          "API_KEY=opaquevalue12345",
-          "TOKEN:opaquevalue12345",
-          agentSecret,
-        ];
-        for (const secret of secretValues) {
-          const config = SAMPLE_CONFIG.replace("route: inference", `route: ${secret}`)
-            .replace("upstream provider: nvidia-prod", `upstream provider: ${secret}`)
-            .replace('default = "backend-dev"', `default = "${agentSecret}"`)
-            .replace('default = "openai:demo-model"', `default = "openai:${secret}"`);
-          const run = runBashWrapper(buildFixture(dir, config), ["status"], {});
-
-          expect(run.status).toBe(0);
-          expect(run.stdout).not.toContain(secret);
-          expect(run.stdout).not.toContain(agentSecret);
-          expect(run.stdout).toContain("Sandbox:  unknown");
-          expect(run.stdout).toContain("Agent:    agent (default)");
-          expect(run.stdout).not.toContain("Route:");
-          expect(run.stdout).not.toContain("Provider:");
-          expect(run.stdout).not.toContain("Model:");
-        }
-      });
-    });
-
-    it("does not write unsafe endpoint values from mutable sources", () => {
-      withTempDir((dir) => {
-        const unsafeEndpoints = [
-          "https://status-user:opaque-password@example.test/v1",
-          "https://example.test/v1?api_key=opaque-secret",
-          "https://example.test/v1#opaque-fragment",
-          "https://status-user:opaque-password\\u0040example.test/v1",
-          "https://example.test/v1\\u003Fapi_key=opaque-secret",
-          "https",
-        ];
-        for (const endpoint of unsafeEndpoints) {
-          for (const source of ["config", "runtime"] as const) {
-            const config =
-              source === "config"
-                ? SAMPLE_CONFIG.replace("https://inference.local/v1", endpoint)
-                : SAMPLE_CONFIG.replace('base_url = "https://inference.local/v1"', "");
-            const env = source === "runtime" ? { OPENAI_BASE_URL: endpoint } : {};
-            const run = runBashWrapper(buildFixture(dir, config), ["status"], env);
-            const refusedByRuntimeGuard = source === "runtime" && /api_key=/i.test(endpoint);
-
-            expect(run.status).toBe(refusedByRuntimeGuard ? 2 : 0);
-            expect(`${run.stdout}\n${run.stderr}`).not.toContain(endpoint);
-            expect(run.stdout).not.toContain("Endpoint:");
-          }
-        }
-      });
-    });
-
-    it("writes safe custom endpoint URLs from the runtime fallback", () => {
-      withTempDir((dir) => {
-        const endpoint = "https://api.example.test:8443/openai/v1";
-        const config = SAMPLE_CONFIG.replace('base_url = "https://inference.local/v1"', "");
-        const run = runBashWrapper(buildFixture(dir, config), ["status"], {
-          OPENAI_BASE_URL: endpoint,
-        });
-
-        expect(run.status).toBe(0);
-        expect(run.stdout).toContain(`Endpoint: ${endpoint}`);
-      });
-    });
-
-    it("advertises the managed identity commands before delegating help upstream", () => {
-      withTempDir((dir) => {
-        const run = runBashWrapper(buildFixture(dir, SAMPLE_CONFIG), ["--help"], {});
-
-        expect(run.status).toBe(0);
-        expect(run.launched).toBe(true);
-        expect(run.stdout).toContain("NemoClaw-managed commands:");
-        expect(run.stdout).toContain("dcode status");
-        expect(run.stdout).toContain("dcode whoami");
-        expect(run.stdout).toContain("dcode identity");
-      });
-    });
-
-    it("reports the sandbox as unknown when the name was not injected", () => {
-      withTempDir((dir) => {
-        const run = runBashWrapper(buildFixture(dir, SAMPLE_CONFIG), ["status"], {});
-
-        expect(run.status).toBe(0);
-        expect(run.launched).toBe(false);
-        expect(run.stdout).toContain("Sandbox:  unknown");
-      });
-    });
-
-    it("still launches dcode for a normal interactive invocation", () => {
-      withTempDir((dir) => {
-        const run = runBashWrapper(buildFixture(dir, SAMPLE_CONFIG), [], {
-          NEMOCLAW_SANDBOX_NAME: "dcode-demo",
-        });
-
-        expect(run.status).toBe(0);
-        expect(run.launched).toBe(true);
-      });
-    });
-  },
-);
-
-describe.skipIf(!canRun)(
-  "agents/langchain-deepagents-code/dcode-wrapper.sh OpenShell infra-key allowlist",
-  () => {
-    it("starts dcode when runtime OPENSHELL_TLS_KEY carries the canonical mounted path", () => {
-      withTempDir((dir) => {
-        const run = runBashWrapper(buildFixture(dir, SAMPLE_CONFIG), ["--version"], {
-          OPENSHELL_TLS_KEY: CANONICAL_TLS_KEY_PATH,
-        });
-
-        expect(run.status).toBe(0);
-        expect(run.launched).toBe(true);
-        expect(run.stderr).not.toContain("refusing to start");
-      });
-    });
-
-    it("refuses noncanonical OpenShell TLS key values without printing them", () => {
-      const pemValue = [
-        "-----BEGIN PRIVATE ",
-        "KEY-----\nraw-private-key\n-----END PRIVATE ",
-        "KEY-----",
-      ].join("");
-      for (const value of [
-        OPAQUE,
-        pemValue,
-        "relative/tls.key",
-        "/tmp/tls.key",
-        `${CANONICAL_TLS_KEY_PATH}.bak`,
-        `tvly-${OPAQUE}`,
-      ]) {
-        withTempDir((dir) => {
-          const run = runBashWrapper(buildFixture(dir, SAMPLE_CONFIG), ["--version"], {
-            OPENSHELL_TLS_KEY: value,
-          });
-
-          expect(run.status).toBe(2);
-          expect(run.launched).toBe(false);
-          expect(run.stderr).toContain("OPENSHELL_TLS_KEY");
-          expect(run.stderr).not.toContain(value);
-        });
-      }
-    });
-
-    it("refuses OpenShell TLS key values in the mutable env file", () => {
-      for (const value of [CANONICAL_TLS_KEY_PATH, OPAQUE]) {
-        withTempDir((dir) => {
-          const fixture = buildFixture(dir, SAMPLE_CONFIG);
-          fs.writeFileSync(fixture.envFile, `OPENSHELL_TLS_KEY=${value}\n`, "utf8");
-
-          const run = runBashWrapper(fixture, ["--version"], {});
-
-          expect(run.status).toBe(2);
-          expect(run.launched).toBe(false);
-          expect(run.stderr).toContain("OPENSHELL_TLS_KEY");
-          expect(run.stderr).toContain(path.join(dir, ".env"));
-          expect(run.stderr).not.toContain(value);
-        });
-      }
-    });
-
-    it("still refuses recognized provider tokens carried by OPENSHELL_TLS_KEY", () => {
-      for (const value of [`nvapi-${OPAQUE}`, `tvly-${OPAQUE}`]) {
-        withTempDir((dir) => {
-          const run = runBashWrapper(buildFixture(dir, SAMPLE_CONFIG), ["--version"], {
-            OPENSHELL_TLS_KEY: value,
-          });
-
-          expect(run.status).toBe(2);
-          expect(run.launched).toBe(false);
-          expect(run.stderr).toContain("OPENSHELL_TLS_KEY");
-          expect(run.stderr).not.toContain(value);
-        });
-      }
-    });
-
-    it("still refuses an opaque credential-name-context variable outside the allowlist", () => {
-      withTempDir((dir) => {
-        const run = runBashWrapper(buildFixture(dir, SAMPLE_CONFIG), [], {
-          CUSTOM_API_KEY: OPAQUE,
-        });
-
-        expect(run.status).toBe(2);
-        expect(run.launched).toBe(false);
-        expect(run.stderr).toContain("CUSTOM_API_KEY");
-      });
-    });
-  },
-);
